requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database & Storage
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0  # PostgreSQL async driver
redis==5.0.1
psycopg2-binary==2.9.9

# AI & LLM
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
tiktoken==0.5.2

# Data Processing & ML
numpy==1.25.2
pandas==2.1.3
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1

# Email Processing
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
email-validator==2.1.0

# Background Tasks & Queuing
celery==5.3.4
flower==2.0.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# HTTP & Networking
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # for testing FastAPI
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# System Monitoring
psutil==5.9.6

# Data Validation & Serialization
marshmallow==3.20.1
cerberus==1.3.5

# Configuration
dynaconf==3.2.4

# Security
cryptography==41.0.8
PyJWT==2.8.0

# Async Support
asyncio-mqtt==0.13.0
aiofiles==23.2.1

# Vector Database (for knowledge base)
faiss-cpu==1.7.4
chromadb==0.4.18

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2

# Rate Limiting
slowapi==0.1.8

# Health Checks & Metrics
aioredis==2.0.1